Plugin SystemInformation ====================== - Correct the MemoryLoad keyword for Linux. - Correct RAM and Virtual Process Memory (divide by 1000) for Windows. - Change historizer Byte and KByte from long long to long.
Plugin WeatherUnderground
========================
- Correct some syntaxes.
- Correct visual unit in debug mode for keyword Temp and WindDirection.

Widget Forecast
==============
- Correct the symbol degree °, to be correctly display under Linux.

+++ sources/plugins/SystemInformation/specificHistorizers/Byte.h
@@ -8,7 +8,7 @@ namespace shared { namespace plugin { namespace yPluginApi { namespace historiza
    //-----------------------------------------------------
    ///\brief A temperature historizable object
    //-----------------------------------------------------
-   class CByte : public CSingleHistorizableData<long long>
+   class CByte : public CSingleHistorizableData<long>
    {
    public:
       //-----------------------------------------------------
